convert matchers to ts
