+++ Commands/destroy_bot.js
@@ -4,7 +4,7 @@ import Game from "../Data/Game.ts";
 import Fixture from "../Data/Fixture.ts";
 import RoomItem from "../Data/RoomItem.ts";
 import Puzzle from "../Data/Puzzle.ts";
-import { itemIdentifierMatches } from "../Modules/matchers.js";
+import { itemIdentifierMatches } from "../Modules/matchers.ts";
 
 /** @import GameSettings from '../Classes/GameSettings.js' */
 /** @import Player from '../Data/Player.ts' */
